refactor(hypr): Rename placements.lua to window_workspaces.lua
